refactor(handler): extract DoUpdate from UpdateWeb
